revmoed redundant code, moved models to apui folder, fixed sris useless ass qr code
