+++ api/sris useless ass qr code.py
