--- myapp/static/css/style.css
@@ -1,3 +1,0 @@
-body {
-    background-color: yellow;
-}
