--- myapp/templates/index.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-{% load static %}
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-    <link rel="stylesheet" type="text/css" href="{% static '/css/style.css'%}">
-</head>
-<body>
-    <script type="text/javascript" src="{% static '/js/index.js' %}"></script>
-    HELLO!!!!
-</body>
-</html>
