--- myapp/urls.py
@@ -1,8 +1,0 @@
-from django.urls import path
-from . import views
-
-
-urlpatterns = [
-    path('', views.indexView),
-    # path("qrcode/", views.scan_qrcode, name="qrcode")
-]
